requirements.txt
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
gradio>=4.0.0
